+++ src/main/java/com/github/fauu/flij/expression/NumberExpression.java
@@ -2,16 +2,16 @@ package com.github.fauu.flij.expression;
 
 import com.github.fauu.flij.reader.lexeme.Lexeme;
 
-public class NumberExpression extends AtomExpression<Float> {
+public class NumberExpression extends AtomExpression<Double> {
   
-  public NumberExpression(float value) {
+  public NumberExpression(double value) {
     super(value);
   }
   
   public NumberExpression(Lexeme lexeme) {
     String rawValue = lexeme.getValue();
 
-    this.value = Float.parseFloat(rawValue);
+    this.value = Double.parseDouble(rawValue);
   }
   
 }
